requirements.txt
aiofiles==0.6.0
asn1crypto==0.24.0
certifi==2018.8.24
chardet==3.0.4
click==7.1.2
colorzero==1.1
cryptography==2.6.1
entrypoints==0.3
Flask==1.1.2
gpiozero==1.5.1
h11==0.12.0
httptools==0.1.1
idna==2.6
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==17.1.1
keyrings.alt==3.1.1
MarkupSafe==1.1.1
psutil==5.8.0
pycairo==1.20.0
pycrypto==2.6.1
pydantic==1.7.3
PyGObject==3.30.4
python-dotenv==0.15.0
pyxdg==0.25
PyYAML==5.3.1
requests==2.21.0
RPi.GPIO==0.7.0
SecretStorage==2.3.1
six==1.12.0
spidev==3.4
ssh-import-id==5.7
starlette==0.13.6
typing-extensions==3.7.4.3
urllib3==1.24.1
uvicorn==0.13.3
uvloop==0.14.0
watchgod==0.6
websockets==8.1
Werkzeug==1.0.1
